==> 01_villes_coordonnees.ipynb <==
import pandas as pd
import requests
import json

villes_coordonees = pd.DataFrame({
    'destination': ["Mont Saint Michel",
"St Malo",
"Bayeux",
"Le Havre",
"Rouen",
"Paris",
"Amiens",
"Lille",
"Strasbourg",
"Chateau du Haut Koenigsbourg",
"Colmar",
"Eguisheim",
"Besancon",
"Dijon",
"Annecy",
"Grenoble",
"Lyon",
"Gorges du Verdon",
"Bormes les Mimosas",
"Cassis",
"Marseille",
"Aix en Provence",
"Avignon",
"Uzes",
"Nimes",
"Aigues Mortes",
"Saintes Maries de la mer",
"Collioure",
"Carcassonne",
"Ariege",
"Toulouse",
"Montauban",
"Biarritz",
"Bayonne",
"La Rochelle"]
})

villes_coordonees['lon'] =0
villes_coordonees['lat'] =0



url = "https://nominatim.openstreetmap.org/search"

params = {
    'q': 'Paris, France',
    'format': 'json',
    'addressdetails': 1,
    'limit': 1
}

headers = {
    'User-Agent': 'maurice.had@gmail.com'
}
response = requests.get(url, params=params, headers=headers)
ville_unique = response.json()
ville_unique

villes_coordonees['lat'] = villes_coordonees['lat'].astype(float)
villes_coordonees['lon'] = villes_coordonees['lon'].astype(float)

lat = float(ville_unique[0]['lat'])
lon = float(ville_unique[0]['lon'])
            
villes_coordonees.at[5, 'lat'] = lat
villes_coordonees.at[5, 'lon'] = lon

villes_coordonees

for index, row in villes_coordonees.iterrows():
    city = row['destination']
    
    params = {
        'q': city,
        'format': 'json',
        'addressdetails': 1
    }
    
    url = "https://nominatim.openstreetmap.org/search"
    headers = {'User-Agent': 'maurice.had@gmail.com'}  
    
    response = requests.get(url, params=params, headers=headers)
    if response.status_code == 200:
        ville_unique = response.json()
        
        lat = float(ville_unique[0]['lat'])
        lon = float(ville_unique[0]['lon'])
            
        villes_coordonees.at[index, 'lat'] = lat
        villes_coordonees.at[index, 'lon'] = lon
        
        print(f'{city} importée')
        
    else:
        print(f"Erreur pour la ville {city} : code {response.status_code}")
        
villes_coordonees

#classé par ordre alphabétique
villes_coordonees = villes_coordonees.sort_values(by='destination').reset_index(drop=True)
villes_coordonees

#changer le nom de la colonne index : 

villes_coordonees = villes_coordonees.reset_index()
villes_coordonees.rename(columns={'index': 'id_destination'}, inplace=True)

villes_coordonees

#exporter en csv :

villes_coordonees.to_csv('./csv_files/villes_coordonnees.csv', index=False)
