==> 04_towns_hotels_urls_booking_to_ csv.ipynb <==
# générer un csv qui concat tous les fichier json x_hotels_urls.json en donnant un id unique à chaque hôtel.

import os
import pandas as pd

# Chemin vers le dossier contenant les fichiers JSON
json_folder = "/Users/maurice/Documents/certification/bloc_1_Build_and_manage_a_data_infrastructure/json_files/hotels_url"

# Liste pour stocker les DataFrames
x_hotels_urls = []

# Parcourir les fichiers JSON dans le dossier
for filename in os.listdir(json_folder):
    file_path = os.path.join(json_folder, filename)
    x_hotels_urls.append(pd.read_json(file_path))

# Concaténer tous les DataFrames
villes_hotels_urls = pd.concat(x_hotels_urls, ignore_index=True)

# Trier les données par "id_destination"
villes_hotels_urls = villes_hotels_urls.sort_values(by="id_destination")

# Ajouter une colonne "id_hotel" avec un identifiant unique pour chaque hôtel
villes_hotels_urls.insert(villes_hotels_urls.columns.get_loc("id_destination") + 1, "id_hotel", range(1, len(villes_hotels_urls) + 1))

# Chemin pour sauvegarder le fichier CSV
output_csv_path = "./csv_files/villes_hotel.csv"

# Sauvegarder les données en CSV
villes_hotels_urls.to_csv(output_csv_path, index=False)
