==> 07_csv_final_villes_meteo_hotel.ipynb <==
# merger les différents csv pour n'en faire qu'un

import os
import pandas as pd

json_hotels_details_path = "/Users/maurice/Documents/certification/bloc_1_Build_and_manage_a_data_infrastructure/json_files/hotels_détails.json"
hotel_csv_path = "/Users/maurice/Documents/certification/bloc_1_Build_and_manage_a_data_infrastructure/csv_files/villes_hotel.csv"
meteo_csv_path = "/Users/maurice/Documents/certification/bloc_1_Build_and_manage_a_data_infrastructure/csv_files/villes_scoring_meteo.csv"
output_csv = "/Users/maurice/Documents/certification/bloc_1_Build_and_manage_a_data_infrastructure/csv_files/final_project.csv"

villes_hotels_df = pd.read_csv(hotel_csv_path)
hotels_details = pd.read_json(json_hotels_details_path)
meteo_df = pd.read_csv(meteo_csv_path)

villes_hotels_df = villes_hotels_df.merge(hotels_details, on="id_hotel", how="left")
villes_hotels_meteo_df = villes_hotels_df.merge(meteo_df, on="id_destination", how="left")

villes_hotels_meteo_df.insert(0, "date_arrivee", villes_hotels_meteo_df.pop("date_arrivee"))
villes_hotels_meteo_df.insert(1, "date_depart", villes_hotels_meteo_df.pop("date_depart"))

villes_hotels_meteo_df.to_csv(output_csv, index=False)
